requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
distro==1.8.0
idna==3.4
JPype1==1.4.1
numpy==1.26.0
packaging==23.1
pandas==2.1.0
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.1
psycopg2-binary==2.9.7
pycparser==2.21
pymssql==2.2.8
pypdfium2==4.20.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.20
tabula-py==2.8.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
